requirements.txt
tensorflow>=2.12.0
keras>=2.12.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.0.0
streamlit>=1.22.0
plotly>=5.13.0
requests>=2.28.0
python-binance>=1.0.17
ta>=0.10.2
seaborn>=0.12.0
matplotlib>=3.7.0
